[outputs] refs #47 - Use accuracy for calculating coins and display sky as float64 instead as uint64.

+++ src/models/outputs/modelOutputs.go
@@ -26,7 +26,7 @@ type QOutput struct {
 	core.QObject
 
 	_ string 	`property:"outputID"`
-	_ uint64 	`property:"addressSky"`
+	_ float64 	`property:"addressSky"`
 	_ uint64 	`property:"addressCoinHours"`
 }
